+++ segway/src/worker/jira.worker.ts
@@ -1,16 +1,12 @@
 //
-import { ConsumeMessage } from "amqplib"
+import { ConsumeMessage } from "amqplib";
 // base worker
-import { BaseWorker } from "../worker";
-
+import { BaseWorker } from "./base.worker";
 
 export class JiraImportWorker extends BaseWorker {
+  constructor() {
+    super("importer");
+  }
 
-    constructor() {
-        super('importer');
-    }
-
-    protected onMessage(msg: ConsumeMessage | null): void {
-    }
-
+  protected onMessage(msg: ConsumeMessage | null): void {}
 }
